Move the 'calculateDaysSince()' function to the boost-score-api js-package.
